Change TaskNumber into short
Byte goes from -127 to 128, we want unsigned value.

+++ core/plugins/org.activitymgr.core.dao/src/org/activitymgr/core/dao/ITaskDAO.java
@@ -83,7 +83,7 @@ public interface ITaskDAO extends IDAO<Task> {
 	 * @throws DAOException
 	 *             levé en cas d'incident technique d'accès à la base.
 	 */
-	byte newTaskNumber(String path) throws DAOException;
+	short newTaskNumber(String path) throws DAOException;
 
 	/**
  	 * Compute the task sums for a given task (when
